mysql/mysql.go
package mysql

import (
	"errors"
	"strings"

	"github.com/mstgnz/sqlporter"
)

type MySQL struct {
	schema *sqlporter.SchemaType
}

// NewMySQL yeni bir MySQL parser instance'ı oluşturur
func NewMySQL() *MySQL {
	return &MySQL{
		schema: &sqlporter.SchemaType{},
	}
}

// Parse MySQL dump'ını parse eder
func (m *MySQL) Parse(content string) (*sqlporter.SchemaType, error) {
	if content == "" {
		return nil, errors.New("boş içerik")
	}

	// TODO: MySQL dump parsing işlemleri burada yapılacak
	return m.schema, nil
}

// Generate verilen şemadan MySQL dump'ı oluşturur
func (m *MySQL) Generate(schema *sqlporter.SchemaType) (string, error) {
	if schema == nil {
		return "", errors.New("boş şema")
	}

	var result strings.Builder

	// TODO: MySQL dump generation işlemleri burada yapılacak

	return result.String(), nil
}
